addded sample data for hooks

+++ breathecode/notify/views.py
@@ -72,6 +72,33 @@ def slack_command(request):
         return Response(str(e), status=status.HTTP_200_OK)
 
 
+@api_view(['GET'])
+@permission_classes([AllowAny])
+def get_sample_data(request):
+
+    items = Hook.objects.filter(user__id=request.user.id)
+    event = request.GET.get('event', None)
+    if event is not None:
+        filtered = True
+        items = items.filter(event__in=event.split(','))
+
+    service_id = request.GET.get('service_id', None)
+    if service_id is not None:
+        filtered = True
+        items = items.filter(service_id__in=service_id.split(','))
+
+    like = request.GET.get('like', None)
+    if like is not None:
+        items = items.filter(Q(event__icontains=like) | Q(target__icontains=like))
+
+    single = items.first()
+    if single is None:
+        return Response({'details': 'No hook found with this filters for sample data'},
+                        status=status.HTTP_400_BAD_REQUEST)
+
+    return Response(single.sample_data)
+
+
 class HooksView(APIView, GenerateLookupsMixin):
     """
     List all snippets, or create a new snippet.
@@ -83,18 +110,16 @@ class HooksView(APIView, GenerateLookupsMixin):
         handler = self.extensions(request)
 
         items = Hook.objects.filter(user__id=request.user.id)
-        lookup = {}
 
-        start = request.GET.get('event', None)
-        if start is not None:
-            start_date = datetime.datetime.strptime(start, '%Y-%m-%d').date()
-            lookup['created_at__gte'] = start_date
+        event = request.GET.get('event', None)
+        if event is not None:
+            filtered = True
+            items = items.filter(event__in=event.split(','))
 
-        if 'event' in self.request.GET:
-            param = self.request.GET.get('event')
-            lookup['event'] = param
-
-        items = items.filter(**lookup)
+        service_id = request.GET.get('service_id', None)
+        if service_id is not None:
+            filtered = True
+            items = items.filter(service_id__in=service_id.split(','))
 
         like = request.GET.get('like', None)
         if like is not None:
@@ -145,7 +170,11 @@ class HooksView(APIView, GenerateLookupsMixin):
                 filtered = True
                 items = items.filter(service_id__in=service_id.split(','))
 
+        if not filtered:
+            raise ValidationException('Please include some filter in the URL')
+
+        total = items.count()
         for item in items:
             item.delete()
 
-        return Response(None, status=status.HTTP_204_NO_CONTENT)
+        return Response({'details': f'Unsubscribed from {total} hooks'}, status=status.HTTP_200_OK)
